# api/openapi.yaml
openapi: 3.1.0
info:
  title: Capcom API
  version: 0.1.0
  description: |
    Runtime-agnostic AgentOps control-plane API.

    All `/v1/*` endpoints require `Authorization: Bearer <CAPCOM_ADMIN_TOKEN>`; only
    `GET /healthz` is unauthenticated. Browser clients on a different origin (the Next.js
    console) are supported via CORS: the allowed origins are configured with
    `CAPCOM_CORS_ALLOWED_ORIGINS` (default `http://localhost:3000,http://127.0.0.1:3000`),
    and preflight `OPTIONS` requests return `204` without requiring admin auth.
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
paths:
...
  /v1/agents/{id}/actions/set-status:
    post:
      summary: Validate or change an agent lifecycle status
      operationId: setAgentStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAgentStatusRequest"
      responses:
        "200":
          description: Audited control action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlAction"
        "400":
          $ref: "#/components/responses/Error"
  /v1/agents/{id}/actions/delete:
    post:
      summary: Validate or delete a runtime agent
      operationId: deleteAgent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAgentRequest"
      responses:
        "200":
          description: Audited control action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlAction"
        "400":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Capcom admin token
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    SetAgentStatusRequest:
      type: object
      required: [status, actor, reason, idempotency_key]
      properties:
        status:
          type: string
          enum: [enabled, disabled]
        actor: { type: string }
        reason: { type: string }
        idempotency_key: { type: string }
        dry_run: { type: boolean, default: false }
    DeleteAgentRequest:
      type: object
      required: [confirmation, actor, reason, idempotency_key]
      properties:
        confirmation: { type: string }
        actor: { type: string }
        reason: { type: string }
        idempotency_key: { type: string }
        dry_run: { type: boolean, default: false }
...
    ControlAction:
      type: object
      required: [id, runtime_connection_id, agent_id, action_type, status, actor, reason, idempotency_key]
      properties:
        id:
          type: string
        runtime_connection_id:
          type: string
        agent_id:
          type: string
        action_type:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed, rejected]
        actor:
          type: string
        reason:
          type: string
        idempotency_key:
          type: string
        result:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
              description: Safe, actionable message suitable for operators.
            retryable:
              type: boolean